Warnings
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@42267 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/RuntimeStatistics.cpp
@@ -262,7 +262,7 @@ PerformanceInfo* RuntimeStatistics::computeDifference(Database* dbb,
 			// Point TraceCounts to counts array from baseline object
 			if (!all_zeros)
 			{
-				jrd_rel* relation = new_cnts->rlc_relation_id < dbb->dbb_relations->count() ?
+				jrd_rel* relation = size_t(new_cnts->rlc_relation_id) < dbb->dbb_relations->count() ?
 					(*dbb->dbb_relations)[new_cnts->rlc_relation_id] : NULL;
 				TraceCounts traceCounts;
 				traceCounts.trc_relation_id = new_cnts->rlc_relation_id;
@@ -276,7 +276,7 @@ PerformanceInfo* RuntimeStatistics::computeDifference(Database* dbb,
 		}
 		else
 		{
-			jrd_rel* relation = new_cnts->rlc_relation_id < dbb->dbb_relations->count() ?
+			jrd_rel* relation = size_t(new_cnts->rlc_relation_id) < dbb->dbb_relations->count() ?
 				(*dbb->dbb_relations)[new_cnts->rlc_relation_id] : NULL;
 
 			// Point TraceCounts to counts array from object with updated counters
